+++ include/pid_controller.hpp
@@ -14,27 +14,27 @@
 class PIDController {
    private:
     // PIDゲイン
-    double kp;  // 比例ゲイン
-    double ki;  // 積分ゲイン
-    double kd;  // 微分ゲイン
+    float kp;  // 比例ゲイン
+    float ki;  // 積分ゲイン
+    float kd;  // 微分ゲイン
 
     // 内部状態
-    double prev_error;  // 前回の偏差
-    double integral;    // 積分値
-    double prev_input;  // 前回の入力値（微分先行型用）
+    float prev_error;  // 前回の偏差
+    float integral;    // 積分値
+    float prev_input;  // 前回の入力値（微分先行型用）
 
     // 制限値
-    double output_min;    // 出力最小値
-    double output_max;    // 出力最大値
-    double integral_min;  // 積分最小値（アンチワインドアップ）
-    double integral_max;  // 積分最大値（アンチワインドアップ）
+    float output_min;    // 出力最小値
+    float output_max;    // 出力最大値
+    float integral_min;  // 積分最小値（アンチワインドアップ）
+    float integral_max;  // 積分最大値（アンチワインドアップ）
 
     // 設定フラグ
     bool use_derivative_on_input;  // 微分先行型フラグ
     bool first_run;                // 初回実行フラグ
 
     // サンプリング時間
-    double dt;  // サンプリング時間[s]
+    float dt;  // サンプリング時間[s]
 
    public:
     /**
@@ -44,7 +44,7 @@ class PIDController {
      * @param kd_val 微分ゲイン
      * @param sample_time サンプリング時間[s]
      */
-    PIDController(double kp_val = 0.0, double ki_val = 0.0, double kd_val = 0.0, double sample_time = 0.001);
+    PIDController(float kp_val = 0.0, float ki_val = 0.0, float kd_val = 0.0, float sample_time = 0.001);
 
     /**
      * @brief PIDゲインを設定
@@ -52,21 +52,21 @@ class PIDController {
      * @param ki_val 積分ゲイン
      * @param kd_val 微分ゲイン
      */
-    void setGains(double kp_val, double ki_val, double kd_val);
+    void setGains(float kp_val, float ki_val, float kd_val);
 
     /**
      * @brief 出力制限を設定
      * @param min_val 最小値
      * @param max_val 最大値
      */
-    void setOutputLimits(double min_val, double max_val);
+    void setOutputLimits(float min_val, float max_val);
 
     /**
      * @brief 積分制限を設定（アンチワインドアップ）
      * @param min_val 積分最小値
      * @param max_val 積分最大値
      */
-    void setIntegralLimits(double min_val, double max_val);
+    void setIntegralLimits(float min_val, float max_val);
 
     /**
      * @brief 微分先行型を有効/無効にする
@@ -78,7 +78,7 @@ class PIDController {
      * @brief サンプリング時間を設定
      * @param sample_time サンプリング時間[s]
      */
-    void setSampleTime(double sample_time);
+    void setSampleTime(float sample_time);
 
     /**
      * @brief PIDコントローラをリセット
@@ -91,7 +91,7 @@ class PIDController {
      * @param input 現在値
      * @return PID出力
      */
-    double compute(double setpoint, double input);
+    float compute(float setpoint, float input);
 
     /**
      * @brief I-P制御計算（積分先行型）
@@ -100,7 +100,7 @@ class PIDController {
      * @param input 現在値
      * @return I-P出力
      */
-    double computeIP(double setpoint, double input);
+    float computeIP(float setpoint, float input);
 
     /**
      * @brief PI-D制御計算（微分先行型）
@@ -108,19 +108,19 @@ class PIDController {
      * @param input 現在値
      * @return PI-D出力
      */
-    double computePID(double setpoint, double input);
+    float computePID(float setpoint, float input);
 
     /**
      * @brief 現在の積分値を取得
      * @return 積分値
      */
-    double getIntegral() const { return integral; }
+    float getIntegral() const { return integral; }
 
     /**
      * @brief 前回の偏差を取得
      * @return 前回偏差
      */
-    double getPrevError() const { return prev_error; }
+    float getPrevError() const { return prev_error; }
 
     /**
      * @brief デバッグ情報を出力
@@ -128,7 +128,7 @@ class PIDController {
      * @param error 現在の偏差
      * @param output 出力値
      */
-    void printDebugInfo(const char* prefix, double error, double output) const;
+    void printDebugInfo(const char* prefix, float error, float output) const;
 };
 
 /**
@@ -144,7 +144,7 @@ class VelocityIPController : public PIDController {
      * @param kp_val 比例ゲイン
      * @param sample_time サンプリング時間[s]
      */
-    VelocityIPController(double ki_val = 0.0, double kp_val = 0.0, double sample_time = 0.001);
+    VelocityIPController(float ki_val = 0.0, float kp_val = 0.0, float sample_time = 0.001);
 
     /**
      * @brief 速度I-P制御計算
@@ -152,7 +152,7 @@ class VelocityIPController : public PIDController {
      * @param current_velocity 現在速度
      * @return 制御出力
      */
-    double computeVelocity(double target_velocity, double current_velocity);
+    float computeVelocity(float target_velocity, float current_velocity);
 };
 
 /**
@@ -169,7 +169,7 @@ class PositionPIDController : public PIDController {
      * @param kd_val 微分ゲイン
      * @param sample_time サンプリング時間[s]
      */
-    PositionPIDController(double kp_val = 0.0, double ki_val = 0.0, double kd_val = 0.0, double sample_time = 0.001);
+    PositionPIDController(float kp_val = 0.0, float ki_val = 0.0, float kd_val = 0.0, float sample_time = 0.001);
 
     /**
      * @brief 位置PID制御計算
@@ -177,7 +177,7 @@ class PositionPIDController : public PIDController {
      * @param current_position 現在位置
      * @return 制御出力（通常は目標速度）
      */
-    double computePosition(double target_position, double current_position);
+    float computePosition(float target_position, float current_position);
 };
 
 #endif  // PID_CONTROLLER_HPP
